fix: parse test suite
